pyproject.toml
[project]
name = "backend-course"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Skvortsov",email = "roman.skwortsoff@mail.ru"}
]
readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.1,<0.35.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
pydantic-settings = ">=2.9.1,<3.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
asyncpg = "^0.30.0"
black = "^25.1.0"
bcrypt = "4.0.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
redis = "^6.1.0"
fastapi-cache2 = "^0.2.2"
celery = "^5.5.2"
pillow = "^11.2.1"
python-multipart = "^0.0.20"
pytest = "^8.3.5"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.26.0"
ruff = "^0.11.12"
confluent-kafka = "^2.10.1"
motor = "^3.7.1"
httpx = "^0.28.1"
aiohttp = "^3.12.13"
elasticsearch = {version = ">=8.11.0,<8.12.0", extras = ["async"]}
nest-asyncio = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88